+++ R/ft_search.R
@@ -54,6 +54,10 @@
 #' # BMC - can be very slow
 #' (res <- ft_search(query='ecology', from='bmc'))
 #' res$bmc
+#' 
+#' # Europe PMC
+#' (res <- ft_search(query='ecology', from='europmc'))
+#' res$europmc
 #'
 #' # PLOS, Crossref, and arxiv
 #' (res <- ft_search(query='ecology', from=c('plos','crossref','arxiv')))
@@ -69,18 +73,23 @@ ft_search <- function(query, from = 'plos', limit = 10,
                       entrezopts = list(),
                       arxivopts = list(),
                       biorxivopts = list(),
+                      euroopts = list(),
                       ...) {
 
-  from <- match.arg(from, c("plos", "bmc", "crossref", "entrez", "arxiv", "biorxiv"), several.ok = TRUE)
+  from <- match.arg(from, 
+                    c("plos", "bmc", "crossref", "entrez", "arxiv", "biorxiv", "europmc"), 
+                    several.ok = TRUE)
   plos_out <- plugin_plos(from, query, limit, plosopts)
   bmc_out <- plugin_bmc(from, query, limit, bmcopts)
   cr_out <- plugin_crossref(from, query, limit, crossrefopts)
   en_out <- plugin_entrez(from, query, limit, entrezopts)
   arx_out <- plugin_arxiv(from, query, limit, arxivopts)
   bio_out <- plugin_biorxivr(from, query, limit, biorxivopts)
+  euro_out <- plugin_europe_pmc(from, query, limit, euroopts)
 
   res <- list(plos = plos_out, bmc = bmc_out, crossref = cr_out,
-              entrez = en_out, arxiv = arx_out, biorxiv = bio_out)
+              entrez = en_out, arxiv = arx_out, biorxiv = bio_out, 
+              europmc = euro_out)
   structure(res, class = "ft", query = query)
 }
 
@@ -98,7 +107,8 @@ print.ft <- function(x, ...) {
     sprintf("Crossref: %s", null_len(x$crossref$found)),
     sprintf("Entrez: %s", null_len(x$entrez$found)),
     sprintf("arxiv: %s", null_len(x$arxiv$found)),
-    sprintf("biorxiv: %s]", null_len(x$biorxiv$found)),
+    sprintf("biorxiv: %s", null_len(x$biorxiv$found)),
+    sprintf("Europe PMC: %s]", null_len(x$europmc$found)),
     sep = "; "), "\n")
 
   cat("Returned:\n")
@@ -108,7 +118,8 @@ print.ft <- function(x, ...) {
     sprintf("Crossref: %s", NROW(x$crossref$data)),
     sprintf("Entrez: %s", NROW(x$entrez$data)),
     sprintf("arxiv: %s", NROW(x$arxiv$data)),
-    sprintf("biorxiv: %s]", NROW(x$biorxiv$data)),
+    sprintf("biorxiv: %s", NROW(x$biorxiv$data)),
+    sprintf("Europe PMC: %s]", NROW(x$europmc$data)),
     sep = "; "), "\n")
 }