+++ R/plugins_search.R
@@ -3,7 +3,7 @@ plugin_plos <- function(sources, query, limit, opts){
     opts$q <- query
     opts$limit <- limit
     out <- do.call(searchplos, opts)
-    zz <- list(found = out$meta$numFound, data = out$data, opts = opts,
+    zz <- list(source = "plos", found = out$meta$numFound, data = out$data, opts = opts,
                license = list(type = "CC-BY",
                               uri = 'http://creativecommons.org/licenses/by/4.0/',
                               text = '<authors> This is an open-access article distributed under
@@ -11,9 +11,10 @@ plugin_plos <- function(sources, query, limit, opts){
                               unrestricted use, distribution, and reproduction in any medium,
                               provided the original author and source are credited.'))
     if (!is(zz$data, "data.frame")) zz$data <- data.frame(NULL)
+    zz <- names_lower(zz)
     structure(zz, class = "ft_ind", query = query)
   } else {
-    zz <- list(found = NULL, data = NULL, opts = opts)
+    zz <- list(source = "plos", found = NULL, data = NULL, opts = opts)
     structure(zz, class = "ft_ind", query = query)
   }
 }
@@ -26,11 +27,12 @@ plugin_crossref <- function(sources, query, limit, opts){
     # opts$filter <- c(`license.url`='http://creativecommons.org/licenses/by/3.0/deed.en_US')
     out <- tryCatch(do.call(cr_works, opts), warning = function(w) w)
     if (is(out, "simpleWarning")) stop(out$message, call. = FALSE)
-    zz <- list(found = out$meta$total_results, data = out$data, opts = opts, 
+    out$data <- names_lower(out$data)
+    zz <- list(source = "crossref", found = out$meta$total_results, data = out$data, opts = opts, 
                license = list(type = "variable, see individual records"))
     structure(zz, class = "ft_ind", query = query)
   } else {
-    zz <- list(found = NULL, data = NULL, opts = opts)
+    zz <- list(source = "crossref", found = NULL, data = NULL, opts = opts)
     structure(zz, class = "ft_ind", query = query)
   }
 }
@@ -41,12 +43,13 @@ plugin_bmc <- function(sources, query, limit, opts){
     opts$limit <- limit
     out <- do.call(bmc_search, opts)
     dat <- do.call(rbind, lapply(out$results$entries, data.frame, stringsAsFactors = FALSE))
+    dat <- names_lower(dat)
     opts$query <- opts$terms; opts$terms <- NULL
-    zz <- list(found = NA, data = dat, opts = opts, 
+    zz <- list(source = "bmc", found = NA, data = dat, opts = opts, 
                license = list(type = "variable, see `isOpenAccess` field in results"))
     structure(zz, class = "ft_ind", query = query)
   } else {
-    zz <- list(found = NULL, data = NULL, opts = opts)
+    zz <- list(source = "bmc", found = NULL, data = NULL, opts = opts)
     structure(zz, class = "ft_ind", query = query)
   }
 }
@@ -71,11 +74,25 @@ plugin_entrez <- function(sources, query, limit, opts){
     row.names(data) <- NULL
     data <- move_col(data, "title")
     data <- move_col(data, "authors")
+    data <- names_lower(data)
 
-    zz <- list(found = as.integer(out$count), data = data, opts = opts)
+    zz <- list(source = "entrez", found = as.integer(out$count), data = data, opts = opts)
     structure(zz, class = "ft_ind", query = query)
   } else {
-    zz <- list(found = NULL, data = NULL, opts = opts)
+    zz <- list(source = "entrez", found = NULL, data = NULL, opts = opts)
+    structure(zz, class = "ft_ind", query = query)
+  }
+}
+
+plugin_europe_pmc <- function(sources, query, limit, opts){
+  if (any(grepl("europmc", sources))) {
+    opts$query <- query
+    out <- do.call(eupmc_search, opts)
+    zz <- list(source = "europmc", found = out$hitCount, 
+               data = out$resultList$result, opts = opts)
+    structure(zz, class = "ft_ind", query = query)
+  } else {
+    zz <- list(source = "europmc", found = NULL, data = NULL, opts = opts)
     structure(zz, class = "ft_ind", query = query)
   }
 }
@@ -85,11 +102,12 @@ plugin_arxiv <- function(sources, query, limit, opts){
     opts$query <- query
     opts$limit <- limit
     out <- do.call(arxiv_search, opts)
-    zz <- list(found = attributes(out)$total_results, data = out, opts = opts,
+    out <- names_lower(out)
+    zz <- list(source = "arxiv", found = attributes(out)$total_results, data = out, opts = opts,
                license = list(type = "variable, but should be free to text-mine, see http://arxiv.org/help/license and http://arxiv.org/help/bulk_data"))
     structure(zz, class = "ft_ind", query = query)
   } else {
-    zz <- list(found = NULL, data = NULL, opts = opts)
+    zz <- list(source = "arxiv", found = NULL, data = NULL, opts = opts)
     structure(zz, class = "ft_ind", query = query)
   }
 }
@@ -99,14 +117,12 @@ plugin_biorxivr <- function(sources, query, limit, opts){
     opts$query <- query
     opts$limit <- limit
     out <- do.call(biorxiv_search, opts)
-    zz <- list(found = out$found, data = out$data, opts = opts, 
+    out$data <- names_lower(out$data)
+    zz <- list(source = "biorxiv", found = out$found, data = out$data, opts = opts, 
                license = list(type = "variable, but free to text-mine, see http://biorxiv.org/about-biorxiv"))
     structure(zz, class = "ft_ind", query = query)
   } else {
-    zz <- list(found = NULL, data = NULL, opts = opts)
+    zz <- list(source = "biorxiv", found = NULL, data = NULL, opts = opts)
     structure(zz, class = "ft_ind", query = query)
   }
-
 }
-
-move_col <- function(x, y) x[ c(names(x)[-grep(y, names(x))], y) ]